Update location of t2t_rl_trainer.
