+++ tensor2tensor/rl/t2t_rl_trainer.py
@@ -15,15 +15,29 @@
 
 """Training of RL agent with PPO algorithm."""
 
-# Dependency imports
+import tensorflow as tf
 
 from tensor2tensor.rl import rl_trainer_lib
 
-import tensorflow as tf
+
+flags = tf.flags
+FLAGS = flags.FLAGS
+
+flags.DEFINE_string("event_dir", None,
+                    "Where to store the event file.")
+flags.DEFINE_string("environment", "pendulum",
+                    "Which environment should be used for training.")
 
 
-def main(_):
-  rl_trainer_lib.train(rl_trainer_lib.pendulum_params())
+def main(argv):
+  name_to_env = {
+      "pendulum": rl_trainer_lib.pendulum_params,
+      "cartpole": rl_trainer_lib.cartpole_params,
+  }
+  if FLAGS.environment not in name_to_env:
+    raise ValueError(
+        "Environment with name %s not configured." % FLAGS.environment)
+  rl_trainer_lib.train(name_to_env[FLAGS.environment](), FLAGS.event_dir)
 
 
 if __name__ == "__main__":
